+++ VS2019/audio_analysis_lib.vcxproj
@@ -26,7 +26,6 @@
     <ClCompile Include="..\src\impl\world_cheaptrick_sp.cpp" />
     <ClCompile Include="..\src\impl\world_d4c_ap.cpp" />
     <ClCompile Include="..\src\impl\world_harvest_f0.cpp" />
-    <ClCompile Include="..\src\impl\world_harvest_f0_impl.cpp" />
     <ClCompile Include="..\src\impl\world_synthesis.cpp" />
     <ClCompile Include="..\src\impl\world_synthesis_impl.cpp" />
   </ItemGroup>
@@ -35,7 +34,6 @@
     <ClInclude Include="..\src\fft_component.hpp" />
     <ClInclude Include="..\src\fft_exception.hpp" />
     <ClInclude Include="..\src\impl\input_capture_impl.hpp" />
-    <ClInclude Include="..\src\impl\world_harvest_f0_impl.hpp" />
     <ClInclude Include="..\src\impl\world_synthesis_impl.hpp" />
     <ClInclude Include="..\src\input_capture.hpp" />
     <ClInclude Include="..\src\lib_audio_analysis.h" />
