+++ VS2019/audio_analysis_lib.vcxproj.filters
@@ -36,9 +36,6 @@
     <ClCompile Include="..\src\impl\world_harvest_f0.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\impl\world_harvest_f0_impl.cpp">
-      <Filter>ソース ファイル</Filter>
-    </ClCompile>
     <ClCompile Include="..\src\impl\world_synthesis.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
@@ -48,9 +45,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\impl\input_capture_impl.hpp">
-      <Filter>ヘッダー ファイル</Filter>
-    </ClInclude>
-    <ClInclude Include="..\src\impl\world_harvest_f0_impl.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
     <ClInclude Include="..\src\impl\world_synthesis_impl.hpp">
