+++ src/impl/world_harvest_f0.cpp
@@ -1,23 +1,61 @@
 #pragma once
 #include "world_harvest_f0.hpp"
+#include "world/harvest.h"
 
-using namespace audio_analysis_lib;
-
-harvest_f0::harvest_f0(int fs, int frame_period)
-	:pimpl(std::make_unique<my_world_voice_feature::harvest_f0_impl>(fs, frame_period))
-{ }
-
-void harvest_f0::f0_estimate(const double* x, int x_length, double* temporal_positions, double* f0)
+namespace audio_analysis_lib
 {
-	pimpl->f0_estimate(x, x_length, temporal_positions, f0);
+	namespace my_world_voice_feature
+	{
+		// impl‚ÌéŒ¾AÀ‘•
+		class harvest_f0_impl
+		{
+		public:
+			harvest_f0_impl(int init_fs, int init_frame_period);
+			void f0_estimate(const double *x, int x_length, double *temporal_positions, double *f0);
+			int get_f0_sample_num(int x_length);
+		private:
+			HarvestOption option;
+			int fs;
+		};
+
+		harvest_f0_impl::harvest_f0_impl(int init_fs, int init_frame_period)
+			: option(), fs(init_fs)
+		{
+			InitializeHarvestOption(&option);
+			option.f0_floor = 40.0;
+			option.frame_period = init_frame_period;
+		}
+
+		void harvest_f0_impl::f0_estimate(const double* x, int x_length, double* temporal_positions, double* f0)
+		{
+			Harvest(x, x_length, fs, &option, temporal_positions, f0);
+		}
+
+		int harvest_f0_impl::get_f0_sample_num(int x_length)
+		{
+			return GetSamplesForHarvest(fs, x_length, option.frame_period);
+		}
+	}
+
+	// harvest‚ÌÀ‘•
+	harvest_f0::harvest_f0(int fs, int frame_period)
+		:pimpl(std::make_unique<my_world_voice_feature::harvest_f0_impl>(fs, frame_period))
+	{ }
+
+	harvest_f0::~harvest_f0() {}
+
+	void harvest_f0::f0_estimate(const double* x, int x_length, double* temporal_positions, double* f0)
+	{
+		pimpl->f0_estimate(x, x_length, temporal_positions, f0);
+	}
+
+	int harvest_f0::get_f0_sample_num(int x_length)
+	{
+		return pimpl->get_f0_sample_num(x_length);
+	}
 }
 
-int harvest_f0::get_f0_sample_num(int x_length)
-{
-	return pimpl->get_f0_sample_num(x_length);
-}
-
-
+// DLLŠÖ”À‘•
 void audio_analysis_lib::create_harvest_f0(int fs, int frame_period, void** func_object)
 {
 	*func_object = new harvest_f0(fs, frame_period);
