+++ src/world_harvest_f0.hpp
@@ -1,16 +1,19 @@
 #pragma once
 #include <memory>
-#include "impl/world_harvest_f0_impl.hpp"
 #include "define_macro.hpp"
 
 namespace audio_analysis_lib
 {
-	class my_world_voice_feature::harvest_f0_impl;
+	namespace my_world_voice_feature
+	{
+		class harvest_f0_impl;
+	}
 
 	class harvest_f0
 	{
 	public:
 		harvest_f0(int init_fs, int init_frame_period);
+		~harvest_f0();
 		void f0_estimate(const double *x, int x_length, double *temporal_positions, double *f0);
 		int get_f0_sample_num(int x_length);
 	private:
